#422: Fix goroutine leak for deleted filesystems.
When a filesystem is deleted, `updateEtcdAboutSnapshots` will block forever
on the `snapshotsModified` channel, which will never send anything as it's
been deleted. However, `updateEtcdAboutSnapshots` runs inside
`runWhileFilesystemLives` which has subscribed to the filesystem death observer,
so an `Observer.Publish` goroutine will block forever trying to notify it that
the filesystem is dead. Result: Two goroutines leaked per deleted filesystem.