#586: Defensive programming - removed a race condition and a goroutine leak in PollTransfer.
Probably won't fix the issue, but was stuff I noticed while looking at the issue.

Race condition: We wrote to the `result` variable from the goroutine and
read it from the main goroutine; that would be OK in the non-timeout case
where a channel synchronisation occurs between the write and the read, but
in the timeout case, `result` is undefined in the main goroutine (but
extensively read from). The fix is to pass the TransferResult *and* the error
back through the channel.

Goroutine leak: In the timeout case, we never read from the channel, and leave
the subordinate goroutine waiting forever.