ops#105: Client-side retry in `dm dot delete` to see if it fixes the smoke tests.
The theory is that the delete in the smoke test is failing because the dot
is not "fully created" on all nodes (due to etcd backlogs); the `dm dot show` loop
only ensures it's present on ONE node.

If so, we can roll the pattern out to other API functions.