FIX: Avoid discarding connection errors, by lazily connecting on demand.
Not all methods on DotmeshAPI need a connection; we originally connected
in the constructor, but discarded errors in order to support cases where
the only methods we wanted to call were ones that would create the correct
state for a connection further down the road.

However, this meant that if we couldn't connect, we'd find out about it
when we got a nil dereference later.

This change lazily connects, so if we never call connection-needing methods,
there's no error to ignore. Therefore, we can report connection errors when
they happen.