#620: Only initialise FSmachines once we've read current state from etcd.
This avoids a problem whereby existing filesystems in ZFS when dotmesh starts
ended up stuck in a "missing" state, because the discovery code attempts to
update the snapshots from the discovered filesystem. However, if the filesystem
isn't in the masters cache because the state from etcd hasn't been loaded yet,
this would fail badly.